Stop attempting to control sk_pacing_rate on an unpatched kernel before Linux 6.0
It simply is not possible to correctly control sk_pacing_rate through
the SO_MAX_PACING_RATE option. The sk_setsockopt() called when setting
the option only lowers but never increases sk_pacing_rate with its
```
    sk->sk_pacing_rate = min(sk->sk_pacing_rate, ulval);
```

As a consequence, we cannot control sk_pacing_rate on an unpatched
kernel before Linux 6.0 in the BPF implementation.