Further prepare requesting hwtstamps from the socket
Not allowed from eBPF code, yet, so only implemented in the module for
now. For the module, there is no straight-forward way to read the
hwtstamp from the current sk_buff, though.

Hardware timestamping must also be requested from the device (driver)
via `ioctl(SIOCSHWTSTAMP)`, this is not implemented here.